refactor out reducer to datatable and streaming copies
