+++ pinot-core/src/main/java/org/apache/pinot/core/query/reduce/ResultReducerFactory.java
@@ -21,10 +21,17 @@ package org.apache.pinot.core.query.reduce;
 import org.apache.pinot.common.request.BrokerRequest;
 import org.apache.pinot.core.query.aggregation.function.AggregationFunction;
 import org.apache.pinot.core.query.aggregation.function.DistinctAggregationFunction;
-import org.apache.pinot.core.query.reduce.streaming.SelectionOnlyStreamingReducer;
+import org.apache.pinot.core.query.reduce.datatable.AggregationDataTableReducer;
+import org.apache.pinot.core.query.reduce.datatable.DataTableReducer;
+import org.apache.pinot.core.query.reduce.datatable.DistinctDataTableReducer;
+import org.apache.pinot.core.query.reduce.datatable.ExplainPlanDataTableReducer;
+import org.apache.pinot.core.query.reduce.datatable.GapFillGroupByDataTableReducer;
+import org.apache.pinot.core.query.reduce.datatable.GroupByDataTableReducer;
+import org.apache.pinot.core.query.reduce.datatable.SelectionDataTableReducer;
 import org.apache.pinot.core.query.reduce.streaming.AggregationDataTableStreamingReducer;
 import org.apache.pinot.core.query.reduce.streaming.DistinctDataTableStreamingReducer;
 import org.apache.pinot.core.query.reduce.streaming.GroupByDataTableStreamingReducer;
+import org.apache.pinot.core.query.reduce.streaming.SelectionOnlyStreamingReducer;
 import org.apache.pinot.core.query.reduce.streaming.StreamingReducer;
 import org.apache.pinot.core.query.request.context.QueryContext;
 import org.apache.pinot.core.util.GapfillUtils;
