+++ pinot-core/src/main/java/org/apache/pinot/core/query/reduce/datatable/DistinctDataTableReducer.java
@@ -16,13 +16,11 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.core.query.reduce;
+package org.apache.pinot.core.query.reduce.datatable;
 
-import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import org.apache.pinot.common.metrics.BrokerMetrics;
@@ -32,25 +30,22 @@ import org.apache.pinot.common.response.broker.SelectionResults;
 import org.apache.pinot.common.utils.DataSchema;
 import org.apache.pinot.common.utils.DataSchema.ColumnDataType;
 import org.apache.pinot.common.utils.DataTable;
-import org.apache.pinot.core.data.table.Record;
 import org.apache.pinot.core.query.aggregation.function.DistinctAggregationFunction;
 import org.apache.pinot.core.query.distinct.DistinctTable;
+import org.apache.pinot.core.query.reduce.DataTableReducerContext;
+import org.apache.pinot.core.query.reduce.DistinctReducerBase;
 import org.apache.pinot.core.query.request.context.QueryContext;
 import org.apache.pinot.core.transport.ServerRoutingInstance;
-import org.apache.pinot.core.util.QueryOptionsUtils;
 
 
 /**
  * Helper class to reduce data tables and set results of distinct query into the BrokerResponseNative
  */
-public class DistinctDataTableReducer implements DataTableReducer {
-  private final DistinctAggregationFunction _distinctAggregationFunction;
-  private final boolean _responseFormatSql;
+public class DistinctDataTableReducer extends DistinctReducerBase implements DataTableReducer {
 
   // TODO: queryOptions.isPreserveType() is ignored for DISTINCT queries.
-  DistinctDataTableReducer(QueryContext queryContext, DistinctAggregationFunction distinctAggregationFunction) {
-    _distinctAggregationFunction = distinctAggregationFunction;
-    _responseFormatSql = QueryOptionsUtils.isResponseFormatSQL(queryContext.getQueryOptions());
+  public DistinctDataTableReducer(QueryContext queryContext, DistinctAggregationFunction distinctAggregationFunction) {
+    super(queryContext, distinctAggregationFunction);
   }
 
   /**
@@ -112,39 +107,5 @@ public class DistinctDataTableReducer implements DataTableReducer {
         brokerResponseNative.setSelectionResults(reduceToSelectionResult(mainDistinctTable));
       }
     }
-  }
-
-  private SelectionResults reduceToSelectionResult(DistinctTable distinctTable) {
-    List<Serializable[]> rows = new ArrayList<>(distinctTable.size());
-    DataSchema dataSchema = distinctTable.getDataSchema();
-    ColumnDataType[] columnDataTypes = dataSchema.getColumnDataTypes();
-    int numColumns = columnDataTypes.length;
-    Iterator<Record> iterator = distinctTable.getFinalResult();
-    while (iterator.hasNext()) {
-      Object[] values = iterator.next().getValues();
-      Serializable[] row = new Serializable[numColumns];
-      for (int i = 0; i < numColumns; i++) {
-        row[i] = columnDataTypes[i].convertAndFormat(values[i]);
-      }
-      rows.add(row);
-    }
-    return new SelectionResults(Arrays.asList(dataSchema.getColumnNames()), rows);
-  }
-
-  private ResultTable reduceToResultTable(DistinctTable distinctTable) {
-    List<Object[]> rows = new ArrayList<>(distinctTable.size());
-    DataSchema dataSchema = distinctTable.getDataSchema();
-    ColumnDataType[] columnDataTypes = dataSchema.getColumnDataTypes();
-    int numColumns = columnDataTypes.length;
-    Iterator<Record> iterator = distinctTable.getFinalResult();
-    while (iterator.hasNext()) {
-      Object[] values = iterator.next().getValues();
-      Object[] row = new Object[numColumns];
-      for (int i = 0; i < numColumns; i++) {
-        row[i] = columnDataTypes[i].convertAndFormat(values[i]);
-      }
-      rows.add(row);
-    }
-    return new ResultTable(dataSchema, rows);
   }
 }
