+++ pinot-core/src/main/java/org/apache/pinot/core/query/reduce/streaming/GroupByDataTableStreamingReducer.java
@@ -21,9 +21,7 @@ package org.apache.pinot.core.query.reduce.streaming;
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
-import java.util.Map;
 import java.util.concurrent.TimeoutException;
-import org.apache.pinot.common.request.context.ExpressionContext;
 import org.apache.pinot.common.request.context.FilterContext;
 import org.apache.pinot.common.response.broker.BrokerResponseNative;
 import org.apache.pinot.common.response.broker.ResultTable;
@@ -33,48 +31,28 @@ import org.apache.pinot.common.utils.DataTable;
 import org.apache.pinot.core.data.table.ConcurrentIndexedTable;
 import org.apache.pinot.core.data.table.IndexedTable;
 import org.apache.pinot.core.data.table.Record;
-import org.apache.pinot.core.query.aggregation.function.AggregationFunction;
 import org.apache.pinot.core.query.reduce.DataTableReducerContext;
+import org.apache.pinot.core.query.reduce.GroupByReducerBase;
 import org.apache.pinot.core.query.reduce.HavingFilterHandler;
 import org.apache.pinot.core.query.reduce.PostAggregationHandler;
 import org.apache.pinot.core.query.request.context.QueryContext;
 import org.apache.pinot.core.transport.ServerRoutingInstance;
 import org.apache.pinot.core.util.GroupByUtils;
-import org.apache.pinot.core.util.QueryOptionsUtils;
 
 
 /**
  * Helper class to reduce data tables and set group by results into the BrokerResponseNative
  */
 @SuppressWarnings({"rawtypes", "unchecked"})
-public class GroupByDataTableStreamingReducer implements StreamingReducer {
+public class GroupByDataTableStreamingReducer extends GroupByReducerBase implements StreamingReducer {
   private static final int MIN_DATA_TABLES_FOR_CONCURRENT_REDUCE = 2; // TBD, find a better value.
-
-  private final QueryContext _queryContext;
-  private final AggregationFunction[] _aggregationFunctions;
-  private final int _numAggregationFunctions;
-  private final List<ExpressionContext> _groupByExpressions;
-  private final int _numGroupByExpressions;
-  private final int _numColumns;
-  private final boolean _preserveType;
-  private final boolean _sqlQuery;
 
   private DataSchema _dataSchema;
   private DataTableReducerContext _reducerContext;
   private IndexedTable _indexedTable;
 
   public GroupByDataTableStreamingReducer(QueryContext queryContext) {
-    _queryContext = queryContext;
-    _aggregationFunctions = queryContext.getAggregationFunctions();
-    assert _aggregationFunctions != null;
-    _numAggregationFunctions = _aggregationFunctions.length;
-    _groupByExpressions = queryContext.getGroupByExpressions();
-    assert _groupByExpressions != null;
-    _numGroupByExpressions = _groupByExpressions.size();
-    _numColumns = _numAggregationFunctions + _numGroupByExpressions;
-    Map<String, String> queryOptions = queryContext.getQueryOptions();
-    _preserveType = QueryOptionsUtils.isPreserveType(queryOptions);
-    _sqlQuery = queryContext.getBrokerRequest().getPinotQuery() != null;
+    super(queryContext);
   }
 
   @Override
@@ -151,29 +129,6 @@ public class GroupByDataTableStreamingReducer implements StreamingReducer {
     } catch (Exception e) {
       throw new RuntimeException(e);
     }
-  }
-
-  /**
-   * Helper method to extract the final aggregation results for the given row (in-place).
-   */
-  private void extractFinalAggregationResults(Object[] row) {
-    for (int i = 0; i < _numAggregationFunctions; i++) {
-      int valueIndex = i + _numGroupByExpressions;
-      row[valueIndex] = _aggregationFunctions[i].extractFinalResult(row[valueIndex]);
-    }
-  }
-
-  /**
-   * Constructs the DataSchema for the rows before the post-aggregation (SQL mode).
-   */
-  private DataSchema getPrePostAggregationDataSchema(DataSchema dataSchema) {
-    String[] columnNames = dataSchema.getColumnNames();
-    ColumnDataType[] columnDataTypes = new ColumnDataType[_numColumns];
-    System.arraycopy(dataSchema.getColumnDataTypes(), 0, columnDataTypes, 0, _numGroupByExpressions);
-    for (int i = 0; i < _numAggregationFunctions; i++) {
-      columnDataTypes[i + _numGroupByExpressions] = _aggregationFunctions[i].getFinalResultColumnType();
-    }
-    return new DataSchema(columnNames, columnDataTypes);
   }
 
   private void appendIndexedTable(DataSchema dataSchema, DataTable dataTable, DataTableReducerContext reducerContext)
