+++ pinot-core/src/test/java/org/apache/pinot/queries/DistinctQueriesTest.java
@@ -42,7 +42,7 @@ import org.apache.pinot.core.data.table.Record;
 import org.apache.pinot.core.operator.BaseOperator;
 import org.apache.pinot.core.operator.blocks.IntermediateResultsBlock;
 import org.apache.pinot.core.query.distinct.DistinctTable;
-import org.apache.pinot.core.query.reduce.BrokerReduceService;
+import org.apache.pinot.core.query.reduce.datatable.BrokerReduceService;
 import org.apache.pinot.core.query.request.context.QueryContext;
 import org.apache.pinot.core.query.request.context.utils.QueryContextConverterUtils;
 import org.apache.pinot.core.transport.ServerRoutingInstance;
